refactor GameScreenDeveloperSidebar
